requirements.txt
Flask==3.0.3
openai==1.79.0
coverage==7.5.1
google-generativeai==0.8.5
prometheus_client==0.19.0
python-json-logger==0.1.11
flask_cors==4.0.0